+++ scene/resources/bit_map.cpp
@@ -332,7 +332,7 @@ Vector<Vector2> BitMap::_march_square(const Rect2i &rect, const Point2i &start) 
 		prevx = stepx;
 		prevy = stepy;
 
-		ERR_FAIL_COND_V(count > width * height, _points);
+		ERR_FAIL_COND_V((int)count > width * height, _points);
 	} while (curx != startx || cury != starty);
 	return _points;
 }
